chore: C++20 is more strict

+++ src/modules/bluefish/producer/bluefish_producer.cpp
@@ -423,11 +423,11 @@ struct bluefish_producer
                 std::lock_guard<std::mutex> lock(state_mutex_);
                 state_["file/name"]              = model_name_;
                 state_["file/path"]              = device_index_;
-                state_["file/video/width"]       = static_cast<long>(width);
-                state_["file/video/height"]      = static_cast<long>(height);
+                state_["file/video/width"]       = width;
+                state_["file/video/height"]      = height;
                 state_["file/audio/sample-rate"] = format_desc_.audio_sample_rate;
                 state_["file/audio/channels"]    = format_desc_.audio_channels;
-                state_["file/fps"]               = static_cast<double>(fps);
+                state_["file/fps"]               = fps;
                 state_["profiler/time"]          = {frame_timer.elapsed(), fps};
                 state_["buffer"]                 = {frame_buffer_.size(), frame_buffer_.capacity()};
             }
